// StockMarket/ClientApp/src/model/guppy.ts
import { TimeSeries } from './base/TimeSeries';

export class guppy extends TimeSeries {

    shortlag3: number;
    shortlag5: number;
    shortlag8: number;
    shortlag10: number;
    shortlag12: number;
    shortlag15: number;
    longlag30: number;
    longlag35: number;
    longlag40: number;
    longlag45: number;
    longlag50: number;
    longlag60: number;

}